requirements.txt
flask
openai
python-dotenv
azure-ai-vision
requests
streamlit
Pillow
